type int -> float

+++ cpp/class/classTest.cpp
@@ -5,19 +5,19 @@ class Person {
     public : string _name;
     public : int _age;
 
-    private : int _koreaScore, _englishScore, _mathScore;
+    private : float _koreaScore, _englishScore, _mathScore;
 
     //이 버전의 컴파일러에서는 생성자 함수를 포함한 모든 멤버함수에게 public을 달아줘야 사용가능한가 보다. default값이 private 인듯
-    public : int getKoreaScore();
-    public : int getEnglishScore();
-    public : int getMathScore();
+    public : float getKoreaScore();
+    public : float getEnglishScore();
+    public : float getMathScore();
 
-    public : void setKoreaScore(int score);
-    public : void setEnglishScore(int score);
-    public : void setMathScore(int score);
+    public : void setKoreaScore(float score);
+    public : void setEnglishScore(float score);
+    public : void setMathScore(float score);
 
     public : Person(); //생성자 함수
-    public : Person(string name, int age, int koreaScore, int englishScore, int mathScore); //생성자 함수 (오버로딩)
+    public : Person(string name, int age, float koreaScore, float englishScore, float mathScore); //생성자 함수 (오버로딩)
 };
 
 Person::Person(){
@@ -29,7 +29,7 @@ Person::Person(){
     _mathScore = 0;
 }
 
-Person::Person(string name, int age = 20, int koreaScore = 0, int englishScore = 0, int mathScore = 0){
+Person::Person(string name, int age = 20, float koreaScore = 0, float englishScore = 0, float mathScore = 0){
     cout << "overloading constructor" << endl;
 
     cout << "overloading constructor" << endl;
@@ -42,27 +42,27 @@ Person::Person(string name, int age = 20, int koreaScore = 0, int englishScore =
     _mathScore = mathScore;
 }
 
-int Person::getKoreaScore(){
+float Person::getKoreaScore(){
     return _koreaScore;
 }
 
-int Person::getEnglishScore(){
+float Person::getEnglishScore(){
     return _englishScore;
 }
 
-int Person::getMathScore(){
+float Person::getMathScore(){
     return _mathScore;
 }
 
-void Person::setKoreaScore(int score){
+void Person::setKoreaScore(float score){
     _koreaScore = score;
 }
 
-void Person::setEnglishScore(int score){
+void Person::setEnglishScore(float score){
     _englishScore = score;
 }
 
-void Person::setMathScore(int score){
+void Person::setMathScore(float score){
     _mathScore = score;
 }
